+++ zarrs/src/array/codec/array_to_array/squeeze.rs
@@ -60,6 +60,7 @@ fn is_identifier_squeeze(identifier: &str) -> bool {
 }
 
 pub(crate) fn create_codec_squeeze(metadata: &MetadataV3) -> Result<Codec, PluginCreateError> {
+    crate::warn_experimental_extension(metadata.name(), "codec");
     let configuration: SqueezeCodecConfiguration = metadata
         .to_configuration()
         .map_err(|_| PluginMetadataInvalidError::new(SQUEEZE, "codec", metadata.to_string()))?;